requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.8.2
scikit-learn==1.5.2
numpy==2.0.2
python-dotenv==1.0.1
streamlit==1.37.1